refactor: split types/index.ts into domain-specific type modules
